parrotdevelopment-challenge-devops-243362a9bea9/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
factory_boy==2.12.0